BJS2-44016 - the task "Grouping users by age"(corrected)
