+++ src/main/java/school/faang/task_44016/User.java
@@ -24,13 +24,12 @@ public class User {
         this.address = address;
     }
 
-    public static Map<Integer, List<User>> groupUsers(List<User> users){
+    public static Map<Integer, List<User>> groupUsers(List<User> users) {
         Map<Integer, List<User>> sameAge = new HashMap<>();
-        for (int cycleForMap = 0; cycleForMap<users.size(); cycleForMap++){
+        for (int cycleForMap = 0; cycleForMap < users.size(); cycleForMap++) {
             List<User> userSameAge = new ArrayList<>();
-
-            for(int cycleForList = 0; cycleForList<users.size(); cycleForList++){
-                if(users.get(cycleForList).age == users.get(cycleForMap).age){
+            for (int cycleForList = 0; cycleForList < users.size(); cycleForList++) {
+                if (users.get(cycleForList).age == users.get(cycleForMap).age) {
                     userSameAge.add(users.get(cycleForList));
                 }
             }
@@ -41,10 +40,17 @@ public class User {
 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
         User user = (User) o;
-        return age == user.age && Objects.equals(name, user.name) && Objects.equals(workplace, user.workplace) && Objects.equals(address, user.address);
+        return age == user.age
+                && Objects.equals(name, user.name)
+                && Objects.equals(workplace, user.workplace)
+                && Objects.equals(address, user.address);
     }
 
     @Override
@@ -54,30 +60,8 @@ public class User {
 
     @Override
     public String toString() {
-        return "User{" +
-                "name='" + name + '\'' +
-                ", AGE=" + age +
-                ", workplace='" + workplace + '\'' +
-                ", address='" + address + '\'' +
-                '}';
+        return "User{" + "name='" + name + '\'' + ", AGE=" + age + ", workplace='"
+                + workplace + '\'' + ", address='" + address + '\'' + '}';
     }
 }
 
-// На всякий случай сохраняю тут main. Не разобралась, где миенно его надо было реализовать
-//        public static void main(String[] args) {
-//        List<User> users = new ArrayList<>();
-//        User alisa = new User("Alisa", 12, " - ", "USA");
-//        User maria = new User("Maria", 15, " - ", "USA");
-//        User alex = new User("Alex", 12, " - ", "USA");
-//        User tom = new User("Tom", 15, " - ", "USA");
-//        User dima = new User("Dima", 15, " - ", "USA");
-//
-//        users.add(alisa);
-//        users.add(maria);
-//        users.add(alex);
-//        users.add(tom);
-//        users.add(dima);
-//
-//        Map<Integer, List<User>> sameAge = groupUsers(users);
-//        System.out.println(sameAge);
-//    }
